Add changes to images
