+++ api/__init__.py
@@ -1,5 +1,5 @@
 from flask import Flask
-from flask_jwt_extended import JWTManager, jwt_required, create_access_token
+from flask_jwt_extended import JWTManager
 
 from config import DevelopmentConfig
